feat: prompt to continue if resources are not found
Affords the user a chance to duck out if resources are not found
during the sanity check. This could occur if a bare Pod were defined,
or some other transient resource (e.g. Jobs which delete themselves)

However, it otherwise would imply an error or drift in your helm
chart from the resources found in cluster